Improve discrimination of mobile users
Currently the break when shrunk to a width below 600px, which happens in splitscreen laptops. This also intentionally breaks the Try it button for tablet users, who will have a poor experience due to the broken IME support. This will not affect touch screen laptop users, but may affect convertible users intentionally.